Update KEmail.php
Sometimes an email letter can not be delivered due to smtp host problems (for instance, with enabled debug mode I received "Error: too many connections" for a few times). So I made a check for a status of every letter.  If a letter hasn't been sent in this time, it will be sent in next time.